% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Zi.R
\name{bsd}
\alias{bsd}
\title{Spline basis derivative}
\usage{
bsd(x, df = NULL, knots = NULL, degree = 3, intercept = FALSE,
  Boundary.knots = range(x))
}
\arguments{
\item{x}{the predictor variable.  Missing values are allowed.}

\item{df}{}

\item{knots}{}

\item{degree}{}

\item{intercept}{}

\item{Boundary.knots}{}
}
\description{
Spline basis derivative.
}

