% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{const_cov}
\alias{const_cov}
\title{Constant shift covariance}
\usage{
const_cov(t, param = c(scale = 1))
}
\arguments{
\item{t}{two-dimensional vector of evaluation points.}

\item{param}{parameter vector consisting of scale parameter tau.}
}
\description{
Constant shift covariance. Note, the resulting covariance matrix is only positive definite if noise is added.
}
\keyword{covariance}

