% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrOptim_inf.R
\name{constrOptim_inf}
\alias{constrOptim_inf}
\title{constrOptim_inf is a modified version of the regular R constrOptim
with the modification described in https://stackoverflow.com/questions/30741117/issue-with-constroptim}
\usage{
constrOptim_inf(theta, f, grad, ui, ci, mu = 1e-04, control = list(),
  method = if (is.null(grad)) "Nelder-Mead" else "BFGS",
  outer.iterations = 100, outer.eps = 1e-05, ..., hessian = FALSE)
}
\description{
constrOptim_inf is a modified version of the regular R constrOptim
with the modification described in https://stackoverflow.com/questions/30741117/issue-with-constroptim
}

