% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{cov_rect}
\alias{cov_rect}
\title{Rectangular evaluation of covariance functions}
\usage{
cov_rect(t, t_p, cov_fct, param, ...)
}
\arguments{
\item{t}{observation points.}

\item{t_p}{new "prediction" points.}

\item{cov_fct}{covariance function.}

\item{...}{arguments passed to cov_fct.}
}
\description{
Generate rectangular evaluations of covariance functions that are typically used for prediction purposes.
}

