% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dv (Autosaved).R, R/dv.R
\name{dv}
\alias{dv}
\title{Derivative of piecewise linear warping function warping}
\usage{
dv(t, tw)

dv(t, tw)
}
\arguments{
\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}
}
\description{
Evaluate the derivative of a piecewise linear warping function

Evaluate the derivative of a piecewise linear warping function
}
\examples{
#TODO
#TODO
}
\keyword{derivative}
\keyword{warp}

