% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate parameters and predict warps}
\usage{
estimate(y, tm, kts, tw, tau, scale, range, smoothness = 2, nouter = 5,
  ninner = 5)
}
\arguments{
\item{y}{list of \eqn{n} functional observations.}

\item{tm}{list of time points corresponding to y.}

\item{tw}{anchor points for the warping parameters.}

\item{tau}{initialization of warp variance.}

\item{scale}{initialization of scale of serially correlated variation.}

\item{range}{initialization of range of serially correlated variation.}
}
\description{
This function does likelihood estimation in the model \deqn{y_i(t)=\theta(v(t, \boldsymbol{w}_i))+x_i(t)+\varepsilon_i(t)} based on iterative local linearization of the model around predictions of the random warping parameters \eqn{\boldsymbol{w}_i}.
}
\examples{
#TODO
}
\keyword{estimation}
\keyword{likelihood}

