% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_growth.R
\name{estimate_growth}
\alias{estimate_growth}
\title{Estimate parameters and predict warps for growth curve data}
\usage{
estimate_growth(y, t, kts, tw, iter = c(5, 5))
}
\arguments{
\item{y}{list of \eqn{n} functional observations. Missing values are allowed.}

\item{t}{list of time points corresponding to y. Should be scaled to have outer endpoints at 0 and 1.}

\item{kts}{a sequence of increasing points specifying the placement of the knots for the basis.}

\item{tw}{anchor points for the warping parameters.}

\item{iter}{two-dimensional numeric consisting of number of outer and inner iterations.}
}
\description{
This function does likelihood estimation in the model \deqn{y_i(t)=\theta(v(t, w_i))+x_i(t)+\epsilon_i(t)} based on iterative local linearization of the model around predictions of the random warping parameters \eqn{w_i}.
}
\seealso{
\code{\link{estimate_generic}} for a more flexible framework.
}
\keyword{estimation}
\keyword{likelihood}

