% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/estimate_smooth_warp.R
\name{estimate_smooth_warp}
\alias{estimate_smooth_warp}
\title{Estimate parameters and predict warps for curve data}
\usage{
estimate_smooth_warp(y, t, amp_cov_par, amp_cov_fct, warp_cov_par, warp_cov_fct,
  kts, tw, intercept = FALSE, iter = c(5, 5), use_warp_gradient = FALSE,
  homeomorphisms = "no", like_optim_control = list())
}
\arguments{
\item{y}{list of \eqn{n} functional observations. Missing values are allowed.}

\item{t}{list of time points corresponding to y. Should be scaled to have outer endpoints at 0 and 1.}

\item{amp_cov_par}{amplitude covariance parameters.}

\item{amp_cov_fct}{amplitude covariance matrix function.}

\item{warp_cov_par}{warp covariance parameters.}

\item{warp_cov_fct}{warp covariance matrix function.}

\item{kts}{anchor points for the B-spline basis used to model \eqn{\theta}.}

\item{tw}{anchor points for the warping parameters.}

\item{iter}{two-dimensional numeric consisting of number of outer and inner iterations.}

\item{use_warp_gradient}{logical. Should warp prediction use gradient based optimization?}

\item{homeomorphisms}{should warps be constrained to be homeomorphisms? Options are: \code{'no'}, \code{'soft'} or \code{'hard'}. 'soft' will project the prediction onto the space of homeomorphisms after each prediction. 'hard' will do the optimiziation in the constrained space (not implemented yet!).}

\item{like_optim_control}{list of control options for likelihood optimization. Parameters are given as \code{c(amp_cov_par, warp_cov_par)} and options include lower, upper, method, ndev (see \code{\link[stats::optim]{optim}}).}
}
\description{
This function does likelihood estimation in the model \deqn{y_i(t)=\theta(v(t, w_i))+x_i(t)+\epsilon_i(t)} based on iterative local linearization of the model around predictions of the random warping parameters \eqn{w_i}.
}
\keyword{estimation}
\keyword{likelihood}

