% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{fill_precision}
\alias{fill_precision}
\title{Construct list of precision matrices}
\usage{
fill_precision(t, cov_fct, param, amp_fct = NULL)
}
\arguments{
\item{t}{list of observation points corresponding to functional samples.}

\item{cov_fct}{covariance function to use.}

\item{param}{parameters for covariance function.}

\item{amp_fct}{functional basis that amplitude variation should be expressed in. Default is \code{NULL}.}
}
\description{
Generates a list of inverse covariance (precision) matrices for the supplied list of observation points.
}

