% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{ind_like}
\alias{ind_like}
\title{Individual locally linearized likelihood function}
\usage{
ind_like(param, sigma_sq, n_par, r, Zi, amp_cov, warp_cov, t, tw)
}
\arguments{
\item{param}{variance parameters.}

\item{sigma_sq}{maximum likelihood estiamte for sigma_sq.}

\item{n_par}{vector consisting of number of variance parameters for each covariance function.}

\item{r}{residual.}

\item{Zi}{Jacobian in the warps of the mean function around the given warp.}

\item{amp_cov}{function for generating amplitude covariance matrix.}

\item{warp_cov}{function for generating warp covariance function}

\item{t}{time variable corresponding to r.}

\item{tw}{anchor points for warp variables.}
}
\description{
Computes the linearized likelihood for a single functional sample given the residual around a given warp and the corresponding Jacobians.
}
\keyword{likelihood}
\keyword{linearization}

