% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/likelihood.R
\name{like}
\alias{like}
\title{Locally linearized likelihood function}
\usage{
like(param, r, Zis, Cinv, smoothness)
}
\arguments{
\item{param}{variance parameters.}

\item{r}{residual.}

\item{Zis}{list of Jacobians in the warps of the mean function around the given warp.}

\item{Cinv}{inverse warp covariance.}
}
\description{
Computes the linearized likelihood given the residual around a given warp and the corresponding Jacobians.
}
\examples{
#TODO
}
\keyword{likelihood}
\keyword{linearization}

