% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{make_cov_fct}
\alias{make_cov_fct}
\title{Generate covariance matrix function from covariance function}
\usage{
make_cov_fct(cov_fct, noise = TRUE, inv_cov_fct = NULL, ...)
}
\arguments{
\item{cov_fct}{covariance function.}

\item{noise}{logical. Should an identity matrix be added to the covariance matrix?}

\item{...}{arguments passed to cov_fct.}
}
\description{
Function that takes a covariance function and returns a function that generates covariance matrices with the given covariance function
}

