% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/warping_functions.R
\name{make_homeo}
\alias{make_homeo}
\title{Project warping parameters onto the set of values that produce homeomorphic warping functions}
\usage{
make_homeo(w, tw, epsilon = 0.2)
}
\arguments{
\item{w}{warping parameters.}

\item{tw}{anchor points for the warping parameters.}

\item{epsilon}{the smallest allowed slope between two points at unit spacing.}
}
\description{
Projects a set of warping parameters onto the set of values that produce homeomorphic warping functions using constrOptim.
}
\keyword{warping}

