% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{Matern_cov}
\alias{Matern_cov}
\title{Generate Matern plus measurement noise covariances}
\usage{
Matern_cov(t, param = c(scale = 1, range = 1, smoothness = 2), noise = TRUE)
}
\arguments{
\item{t}{evaluation points.}

\item{param}{parameter vector consisting of scale, range and smoothness.}

\item{noise}{logical, should a diagonal matrix be added to the Matern covariance?}
}
\description{
This function generates a Matern motion covariance matrix corresponding to specified evaluation points.
}
\examples{
t <- seq(0, 1, length = 10)
Matern_cov(t, param = c(1, 1, 1/2))
}
\keyword{covariance}

