% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pavpop_clustering.R
\name{pavpoc}
\alias{pavpoc}
\title{Estimate parameters and predict warps for curve data}
\usage{
pavpoc(y, t, basis_fct, warp_fct, amp_cov = NULL, warp_cov = NULL,
  cluster = list(n_clusters = 1, weights = NULL), iter = c(5, 5, 10),
  parallel = list(n_cores = 1, parallel_likelihood = FALSE),
  use_warp_gradient = FALSE, homeomorphisms = "no",
  like_optim_control = list())
}
\arguments{
\item{y}{list of \eqn{n} functional observations. Missing values are allowed.}

\item{t}{list of time points corresponding to y. Should be scaled to have outer endpoints at 0 and 1.}

\item{basis_fct}{basis function to describe the mean function.}

\item{iter}{two-dimensional numeric consisting of number of outer and inner iterations.}

\item{use_warp_gradient}{logical. Should warp prediction use the exact gradient for based optimization?}

\item{homeomorphisms}{should warps be constrained to be homeomorphisms? Options are: \code{'no'}, \code{'soft'} or \code{'hard'}. 'soft' will project the prediction onto the space of homeomorphisms after each prediction. 'hard' will do the optimiziation in the constrained space (not implemented yet!).}

\item{like_optim_control}{list of control options for likelihood optimization. Parameters are given as \code{c(amp_cov_par, warp_cov_par)} and options include lower, upper, method, ndev (see \code{\link[stats::optim]{optim}}).}

\item{amp_cov_par}{amplitude covariance parameters.}

\item{amp_cov_fct}{amplitude covariance matrix function.}

\item{warp_cov_par}{warp covariance parameters.}

\item{warp_cov_fct}{warp covariance matrix function.}

\item{tw}{anchor points for the warping parameters.}

\item{smooth_warp}{logical. Should warping functions be based on a monotonic cubic spline?}

\item{n_cores}{number of cores to use.}
}
\description{
This function does likelihood estimation in the model \deqn{y_i(t)=\theta(v(t, w_i))+x_i(t)+\epsilon_i(t)} based on iterative local linearization of the model around predictions of the random warping parameters \eqn{w_i}.
}
\keyword{estimation}
\keyword{likelihood}

