% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavpop.R
\docType{package}
\name{pavpop-package}
\alias{pavpop-package}
\title{pavpop: A package for simultaneously analyzing functional data
containing systematic phase and amplitude variation.}
\description{
pavpop: A package for simultaneously analyzing functional data
containing systematic phase and amplitude variation.
}
\author{
Lars Lau Raket \email{larslau@math.ku.dk}
}
\references{
L.L. Raket, S. Sommer, and B. Markussen, "A nonlinear mixed-effects model for simultaneous smoothing and registration of functional data," Pattern Recognition Letters, vol. 38, pp. 1-7, 2014, 10.1016/j.patrec.2013.10.018.
}

