% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavpop_amp.R
\name{pavpop_amp}
\alias{pavpop_amp}
\title{Estimate parameters and predict warps for curve data}
\usage{
pavpop_amp(y, t, basis_fct, amp_fct, warp_fct, amp_cov = NULL,
  warp_cov = NULL, warped_amp = FALSE, iter = c(5, 5),
  parallel = list(n_cores = 1, parallel_likelihood = FALSE),
  use_warp_gradient = FALSE, warp_optim_method = "CG",
  homeomorphisms = "no", like_optim_control = list())
}
\arguments{
\item{y}{list of \eqn{n} functional observations. Missing values are allowed.}

\item{t}{list of time points corresponding to y. Should be scaled to have outer endpoints at 0 and 1.}

\item{basis_fct}{basis function to describe the mean function.}

\item{amp_cov}{amplitude covariance matrix function. If `NULL`, the amplitude is assumed to only contain iid Gaussian noise.}

\item{warp_cov}{warp covariance matrix function. If `NULL` the warps are treated as fixed parameters.}

\item{warped_amp}{logical. Does the amplitude correlation follow observed or warped time?}

\item{iter}{two-dimensional numeric consisting of number of outer and inner iterations.}

\item{use_warp_gradient}{logical. Should warp prediction use the exact gradient for based optimization?}

\item{homeomorphisms}{should warps be constrained to be homeomorphisms? Options are: \code{'no'}, \code{'soft'} or \code{'hard'}. 'soft' will project the prediction onto the space of homeomorphisms after each prediction. 'hard' will do the optimiziation in the constrained space (not implemented yet!).}

\item{like_optim_control}{list of control options for likelihood optimization. Parameters are given as \code{c(amp_cov_par, warp_cov_par)} and options include lower, upper, method, ndev (see \code{\link[stats::optim]{optim}}).}

\item{smooth_warp}{logical. Should warping functions be based on a monotonic cubic spline?}

\item{n_cores}{number of cores to use.}

\item{basis_amp}{logical. Should the amplitude variation be expressed in terms of the basis functions?}
}
\description{
This function does likelihood estimation in the model \deqn{y_i(t)=\theta(v(t, w_i))+x_i(t)+\epsilon_i(t)} based on iterative local linearization of the model around predictions of the random warping parameters \eqn{w_i}.
}
\keyword{estimation}
\keyword{likelihood}

