% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Posterior of the data given the random warping parameters}
\usage{
posterior(w, t, y, tw, c, Ainv, tau, Cinv, kts, Boundary.knots = c(0, 1))
}
\arguments{
\item{w}{warping parameters.}

\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{c}{B-spline coefficients.}

\item{Ainv}{precision matrix for amplitude variation.}

\item{tau}{variance parameter (scale) for the warping parameters.}

\item{Cinv}{precision matrix for the warping parameters.}

\item{Boundary.knots}{boundary knots for the B-spline basis.}
}
\description{
This function calculates the posterior of the data given the random warping parameters
}
\examples{
TODO
}
\keyword{posterior}
\keyword{warping}

