% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{posterior_grad}
\alias{posterior_grad}
\title{Posterior of the data given the random warping parameters}
\usage{
posterior_grad(w, dwarp, t, y, tw, c, Ainv, Cinv, kts)
}
\arguments{
\item{dwarp}{Jaobian of the vector of observed warped points in the warping parameters.}

\item{t}{array of time variables corresponding to r.}

\item{tw}{anchor points for warp variables.}
}
\description{
This function calculates the posterior of the data given the random warping parameters
}
\keyword{posterior}
\keyword{warping}

