% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_curve_verbose}
\alias{predict_curve_verbose}
\title{Predict individual curves}
\usage{
predict_curve_verbose(t_p, t, y, c, basis_fct, warp_fct, amp_cov, amp_cov_par,
  w = NULL, warp_cov = NULL, warp_cov_par = NULL, deriv = FALSE,
  force_increasing = FALSE)
}
\arguments{
\item{t_p}{evaluation points for prediction.}

\item{t}{observation points.}

\item{y}{observation values.}

\item{c}{weights for basis functions.}

\item{basis_fct}{basis function to describe the mean function.}

\item{amp_cov_par}{amplitude covariance parameters.}

\item{w}{warping parameters.}

\item{deriv}{logical. Should the derivative of the curve be predicted?}

\item{amp_cov_fct}{amplitude covariance matrix function.}

\item{tw}{anchor points for the warping parameters.}
}
\value{
list with predicted curve, warp parameters, and additional information
}
\description{
This function predicts individual curves from estimated parameters and warps and returns verbose information about the prediction
}
\keyword{curve}
\keyword{prediction}

