% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{predict_warp_pyramid}
\alias{predict_warp_pyramid}
\title{Minimize posterior of the data given the random warping parameters using a pyramidal primal-dual approach}
\usage{
predict_warp_pyramid(w, y, Ainv, t, tw, kts, warp_cov_par, warp_cov_fct,
  plevels = 5, beta0 = 1, start = 1, homeomorphic = TRUE, iter = c(10,
  5), intercept = FALSE)
}
\arguments{
\item{w}{warping parameters.}

\item{Ainv}{precision matrix for amplitude variation.}

\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{c}{B-spline coefficients.}

\item{Cinv}{precision matrix for the warping parameters.}

\item{Boundary.knots}{boundary knots for the B-spline basis.}
}
\description{
This function minimizes the posterior of the data given the random warping parameters using a primal-dual approach in a pyramidal setup
}
\keyword{posterior}
\keyword{warping}

