% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{sigmasq}
\alias{sigmasq}
\alias{sigmasq_amp}
\title{Variance scale estimate from locally linearized likelihood function}
\usage{
sigmasq(param, n_par, r, Zis, amp_cov, warp_cov, t, tw)

sigmasq_amp(param, n_par, r, Zis, amp_cov, warp_cov, amp_fct, t, tw)
}
\arguments{
\item{param}{variance parameters.}

\item{n_par}{vector consisting of number of variance parameters for each covariance function.}

\item{r}{residual.}

\item{Zis}{list of Jacobians in the warps of the mean function around the given warp.}

\item{amp_cov}{function for generating amplitude covariance matrix.}

\item{warp_cov}{function for generating warp covariance function}

\item{t}{array of time variables corresponding to r.}

\item{tw}{anchor points for warp variables.}

\item{amp_fct}{functional basis that amplitude variation should be expressed in. If used,
it is automatically assumed that iid. Gaussian noise is present in the data.}
}
\description{
Estimates the common variance scale of all random effects in the model
from the linearized likelihood given the residual around a predicted
warp and the corresponding Jacobians.
}
\section{Functions}{
\itemize{
\item \code{sigmasq_amp}: Variance scale estimate from locally linearized
likelihood function with amplitude variation described in a functional basis.
}}
\keyword{likelihood}
\keyword{linearization}

