% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/spline_weights.R
\name{spline_weights}
\alias{spline_weights}
\title{Estimate spline weights}
\usage{
spline_weights(y, tm, w, tw, Ainv, kts)
}
\arguments{
\item{y}{list of \eqn{n} functional observations.}

\item{tm}{list of time points corresponding to y.}

\item{w}{warping parameters.}

\item{tw}{anchor points for the warping parameters.}

\item{Ainv}{global precision matrix for amplitude variation.}

\item{kts}{B-spline knots.}
}
\description{
Estimate spline weights for given warps and covariances
}
\examples{
t <- seq(0, 1, length = 10)
plot(v(0.4, t, 0.5), type = 'b', pch = 19)
}
\keyword{warping}

