% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{zero_cov}
\alias{zero_cov}
\title{Zero covariance function}
\usage{
zero_cov(t, param = NULL)
}
\arguments{
\item{t}{evaluation points.}

\item{param}{not used.}
}
\description{
Returns a matrix with zeros with rows and columns equal to the length of t.
}
\keyword{covariance}

