% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zi.R
\name{Zi}
\alias{Zi}
\title{Jacobian of warped B-spline function}
\usage{
Zi(t, dwarp, basis_fct, c)
}
\arguments{
\item{t}{(warped) evaluation points.}

\item{dwarp}{Jacobian of the warping function for the given evaluation points.}

\item{basis_fct}{basis function.}

\item{c}{spline weights.}
}
\description{
Computes the Jacobian of a warped B-spline function.
}

