% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{Brownian_cov}
\alias{Brownian_cov}
\title{Generate Brownian covariances}
\usage{
Brownian_cov(t, tau, type = "motion")
}
\arguments{
\item{t}{evaluation points.}

\item{tau}{scale parameter.}

\item{type}{type of covariance, either 'motion' or 'bridge'.}
}
\description{
This function generates a Brownian motion/bridge covariance matrix corresponding to specified evaluation points.
}
\examples{
t <- seq(0, 1, length = 10)
Brown_cov(t, 1)
}
\keyword{covariance}

