% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/covariances.R
\name{Brownian_motion_cov_fast}
\alias{Brownian_motion_cov_fast}
\title{Generate Brownian motion covariances}
\usage{
Brownian_motion_cov_fast(t)
}
\arguments{
\item{t}{evaluation points.}
}
\description{
This function generates a Brownian motion covariance matrix corresponding to specified evaluation points.
}
\examples{
t <- seq(0, 1, length = 10)
Brown_motion_cov_fast(t)
}
\keyword{covariance}

