% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/warping_functions.R
\name{dv}
\alias{dv}
\title{Derivative of piecewise linear warping function warping}
\usage{
dv(t, tw)
}
\arguments{
\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}
}
\description{
Evaluate the derivative of a piecewise linear warping function
}
\keyword{derivative}
\keyword{warp}

