% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{id_cov}
\alias{id_cov}
\title{Identity covariance function}
\usage{
id_cov(t, param = c(scale = 1))
}
\arguments{
\item{t}{evaluation points.}

\item{param}{either one-dimensional or equal to the length of t consisting of the diagonal entries.}
}
\description{
Returns a diagonal matrix with diagonal given by param.
}
\keyword{covariance}

