% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{id_cov}
\alias{id_cov}
\title{Identity covariance function}
\usage{
id_cov(t, param = c(scale = 1))
}
\arguments{
\item{t}{evaluation points.}

\item{param}{one-dimensional consisting of the diagonal entry.}
}
\description{
Returns a diagonal matrix with diagonal given by param.
}
\keyword{covariance}

