% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{ispline}
\alias{ispline}
\title{Generate increasing spline basis}
\usage{
ispline(x, knots, d)
}
\arguments{
\item{x}{predictor variable.}

\item{knots}{the internal breakpoints of the spline.}

\item{d}{order of the spline functions.}
}
\description{
Method for generating a 'basis function' function
}
\note{
Method is a corrected version of the increasing spline basis code in the \code{SVMMaj} package.
}
\keyword{spline}

