% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{like}
\alias{like}
\title{Locally linearized likelihood function}
\usage{
like(param, n_par, r, Zis, amp_cov, warp_cov, t, tw)
}
\arguments{
\item{param}{variance parameters.}

\item{n_par}{vector consisting of number of variance parameters for each covariance function.}

\item{r}{residual.}

\item{Zis}{list of Jacobians in the warps of the mean function around the given warp.}

\item{amp_cov}{function for generating amplitude covariance matrix.}

\item{warp_cov}{function for generating warp covariance function}

\item{t}{array of time variables corresponding to r.}

\item{tw}{anchor points for warp variables.}

\item{observation_weights}{vector of weights for the individual functional samples to be applied to the likelihood. This is useful for clustering analysis.}
}
\description{
Computes the linearized likelihood given the residual around a given warp and the corresponding Jacobians.
}
\keyword{likelihood}
\keyword{linearization}

