% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{make_basis_fct}
\alias{make_basis_fct}
\title{Generate basis function}
\usage{
make_basis_fct(kts, intercept = FALSE, increasing = FALSE,
  only_intercept = FALSE, order = 3, boundary = c(0, 1))
}
\arguments{
\item{kts}{a sequence of increasing points specifying the placement of the knots.}

\item{intercept}{logical. Should the basis include an intercept?}

\item{increasing}{logical. Should the basis be an I-spline?}

\item{order}{the order of the basis splines.}

\item{boundary}{boundary knots.}
}
\description{
Method for generating a 'basis function' function
}
\keyword{spline}

