% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warping_functions.R
\name{make_warp_fct}
\alias{make_warp_fct}
\title{Generate a warping function}
\usage{
make_warp_fct(type = c("shift", "linear", "piecewise-linear", "smooth"),
  tw = NULL, fix_right = TRUE)
}
\arguments{
\item{type}{type of warping function.}

\item{tw}{anchor points for the warping parameters. Needed for types 'piecewise-linear' and 'smooth'.}
}
\description{
This function generates a certain kind of warping function that can be used by pavpop.
}
\keyword{warping}

