% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{Matern}
\alias{Matern}
\title{Matern covariance function}
\usage{
Matern(d, param = c(scale = 1, range = 1, smoothness = 2))
}
\arguments{
\item{d}{evaluation points.}

\item{param}{parameter vector consisting of scale, range and smoothness.}
}
\description{
Functional form of Matern covariance function. Code adapted from the fields package.
}
\note{
Method taken from the \code{fields} package.
}
\examples{
Matern(seq(0, 1, length = 10), param = c(scale = 1, range = 0.5, smoothness = 2))
Matern(seq(0, 1, length = 10), param = c(scale = 1, range = 1, smoothness = 2))
}
\keyword{covariance}

