% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/covariances.R
\name{Matern_cov}
\alias{Matern_cov}
\title{Generate Matern covariances}
\usage{
Matern_cov(t, scale, range, smoothness)
}
\arguments{
\item{t}{evaluation points.}

\item{scale}{scale parameter.}

\item{range}{range parameter.}

\item{smoothness}{smoothness parameter.}
}
\description{
This function generates a Matern motion covariance matrix corresponding to specified evaluation points.
}
\examples{
t <- seq(0, 1, length = 10)
Matern_cov(t, 1, 1, 1/2)
}
\keyword{covariance}

