% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{posterior}
\alias{posterior}
\title{Posterior of the data given the random warping parameters}
\usage{
posterior(w, t, y, tw, c, Ainv, Cinv, kts)
}
\arguments{
\item{w}{warping parameters.}

\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{c}{B-spline coefficients.}

\item{Ainv}{precision matrix for amplitude variation.}

\item{Cinv}{precision matrix for the warping parameters.}

\item{kts}{anchor points for the B-spline basis used to model the functional parameter \eqn{\theta}.}
}
\description{
This function calculates the posterior of the data given the random warping parameters
}
\keyword{posterior}
\keyword{warping}

