% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{posterior_grad}
\alias{posterior_grad}
\title{Posterior of the data given the random warping parameters}
\usage{
posterior_grad(w, dwarp, t, y, tw, c, Ainv, Cinv, basis_fct)
}
\arguments{
\item{w}{warping parameters.}

\item{dwarp}{Jaobian of the vector of observed warped points in the warping parameters.}

\item{t}{evaluation points.}

\item{y}{values at evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{c}{B-spline coefficients.}

\item{Ainv}{precision matrix for amplitude variation.}

\item{Cinv}{precision matrix for the warping parameters.}

\item{basis_fct}{basis function to describe the mean function.}

\item{smooth_warp}{logical. Should warping functions be based on a monotonic cubic spline?}
}
\description{
This function calculates the posterior of the data given the random warping parameters
}
\keyword{posterior}
\keyword{warping}

