% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_curve}
\alias{predict_curve}
\title{Predict individual curves}
\usage{
predict_curve(t_p, t, y, c, basis_fct, warp_fct, amp_cov, amp_cov_par, w,
  deriv = FALSE)
}
\arguments{
\item{t_p}{evaluation points for prediction.}

\item{t}{observation points.}

\item{y}{observation values.}

\item{c}{weights for basis functions.}

\item{basis_fct}{basis function to describe the mean function.}

\item{amp_cov_par}{amplitude covariance parameters.}

\item{w}{warping parameters.}

\item{deriv}{logical. Should the derivative of the curve be predicted?}

\item{amp_cov_fct}{amplitude covariance matrix function.}

\item{tw}{anchor points for the warping parameters.}
}
\description{
This function predicts individual curves from estimated parameters and warps.
}
\keyword{curve}
\keyword{prediction}

