% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{sigmasq}
\alias{sigmasq}
\title{Noise scale estimate from locally linearized likelihood function}
\usage{
sigmasq(param, n_par, r, Zis, amp_cov_fct, warp_cov_fct, t, tw)
}
\arguments{
\item{param}{variance parameters.}

\item{n_par}{vector consisting of number of variance parameters for each covariance function.}

\item{r}{residual.}

\item{Zis}{list of Jacobians in the warps of the mean function around the given warp.}

\item{amp_cov_fct}{function for generating amplitude covariance matrix.}

\item{warp_cov_fct}{function for generating warp covariance function}

\item{t}{array of time variables corresponding to r.}

\item{tw}{anchor points for warp variables.}
}
\description{
Computes the noise scale estimate from the linearized likelihood given the residual around a predicted warp and the corresponding Jacobians.
}
\keyword{likelihood}
\keyword{linearization}

