% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/spline_weights.R
\name{spline_weights}
\alias{spline_weights}
\title{Estimate spline weights}
\usage{
spline_weights(y, t, w, tw, Ainv, kts, intercept = FALSE,
  smooth_warp = FALSE)
}
\arguments{
\item{y}{list of \eqn{n} functional observations.}

\item{t}{list of time points corresponding to y.}

\item{w}{warping parameters.}

\item{tw}{anchor points for the warping parameters.}

\item{Ainv}{global precision matrix for amplitude variation.}

\item{kts}{B-spline knots.}

\item{smooth_warp}{logical. Should a smooth warping function be used?}
}
\description{
Estimate spline weights for given warps and covariances
}
\keyword{warping}

