% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline_weights.R
\name{spline_weights}
\alias{spline_weights}
\title{Estimate spline weights}
\usage{
spline_weights(y, t, warp_fct, w, Sinv, basis_fct, weights = NULL)
}
\arguments{
\item{y}{list of \eqn{n} functional observations.}

\item{t}{list of time points corresponding to y.}

\item{warp_fct}{warping function.}

\item{w}{warping parameters.}

\item{Sinv}{global precision matrix for amplitude variation.}

\item{basis_fct}{function for generating a basis.}

\item{weights}{weights for the individual observations}
}
\description{
Estimate spline weights for given warps and covariances
}
\keyword{basis}
\keyword{spline}

