% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/warping_functions.R
\name{v}
\alias{v}
\title{Evaluate warping function given weights}
\usage{
v(w, t, tw, smooth = FALSE)
}
\arguments{
\item{w}{warping parameters.}

\item{t}{evaluation points.}

\item{tw}{anchor points for the warping parameters.}

\item{smooth}{logical. Should warping function be based on a monotonic cubic spline?}
}
\description{
This function evaluates the warping function corresponding to a specified set of warping parameters at a set of evaluation points using linear interpolation.
}
\examples{
t <- seq(0, 1, length = 10)
plot(t, v(0.4, t, 0.5), type = 'l', pch = 19)
}
\keyword{warping}

