% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/Zi.R
\name{Zi}
\alias{Zi}
\title{Jacobian of warped B-spline function}
\usage{
Zi(t, dwarp, c, kts)
}
\arguments{
\item{t}{(warped) evaluation points.}

\item{kts}{B-spline knots.}

\item{dvt}{Jacobian of the warping function for the given evaluation points.}
}
\description{
Computes the Jacobian of a warped B-spline function.
}
\examples{
#TODO
}

