% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Zi.R
\name{Zi}
\alias{Zi}
\title{Jacobian of warped B-spline function}
\usage{
Zi(t, dwarp, c, basis_fct)
}
\arguments{
\item{t}{(warped) evaluation points.}

\item{dwarp}{Jacobian of the warping function for the given evaluation points.}

\item{c}{spline weights.}

\item{basis_fct}{basis function.}
}
\description{
Computes the Jacobian of a warped B-spline function.
}

